restructure file tree
