+++ src/tests/counterSlice.spec.js
